% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLWebServiceFunctions.R
\name{DSLWebServiceInterpretation}
\alias{DSLWebServiceInterpretation}
\title{DSL interpretation}
\usage{
DSLWebServiceInterpretation(command, ...)
}
\arguments{
\item{command}{The command to be interpreted.}

\item{...}{Arguments for \code{\link{DSLInterpretationURL}}.}
}
\description{
Gives interpreter web service URL of a DSL command.
}
