% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-ECMMon.R
\name{ECMMonInterpret}
\alias{ECMMonInterpret}
\title{Interpret ECMMon natural speech command.}
\usage{
ECMMonInterpret(command)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses and interprets a command by directly invoking a pair of Raku Perl6 parser and actions classes.
}
\seealso{
Other Spoken ECMMon: 
\code{\link{ECMMonParse}()},
\code{\link{to_ECMMon_R_command}()}
}
\concept{Spoken ECMMon}
