% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-ECMMon.R
\name{ECMMonParse}
\alias{ECMMonParse}
\title{Parse ECMMon natural speech command.}
\usage{
ECMMonParse(command)
}
\arguments{
\item{command}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken ECMMon: 
\code{\link{ECMMonInterpret}()},
\code{\link{to_ECMMon_R_command}()}
}
\concept{Spoken ECMMon}
