% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLWebServiceFunctions.R
\name{InterpretByDSLWebService}
\alias{InterpretByDSLWebService}
\title{Interpret by DSL web service}
\usage{
InterpretByDSLWebService(
  command,
  scheme = "http",
  hostname = "accendodata.net",
  port = "5040",
  path = "translate/",
  url = NULL,
  sub = NULL
)
}
\arguments{
\item{command}{The command to be interpreted.}

\item{scheme}{Scheme, string.}

\item{hostname}{Hostname, string.}

\item{port}{Port, string or integer.}

\item{path}{Path, string.}

\item{url}{URL, string.
If NULL then the URL is composed with the arguments \code{scheme, hostname, port, path}.}

\item{sub}{Sub-service name.}
}
\value{
Returns a list of the form \code{list( Success = <logical>, Response = <httr::GET result>, Content = <content>)}.
}
\description{
Gives interpreter web service URL of a DSL command.
}
\details{

}
\seealso{
Other DSLWebService: 
\code{\link{DSLWebServiceInterpretationURL}()},
\code{\link{DSLWebServiceInterpretation}()},
\code{\link{MakeDSLWebServiceURL}()}
}
\concept{DSLWebService}
