% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Parsing-and-interpretation-functions-hook-up-for-spoken-dplyr.R
\docType{data}
\name{localUserDirName}
\alias{localUserDirName}
\title{Local user directory name}
\format{An object of class \code{character} of length 1.}
\usage{
localUserDirName
}
\description{
Local user directory name
}
\keyword{datasets}
