% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-LSAMon.R
\name{LSAMonInterpret}
\alias{LSAMonInterpret}
\title{Interpret LSAMon natural speech command.}
\usage{
LSAMonInterpret(command)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses and interprets a command by directly invoking a pair of Raku Perl6 parser and actions classes.
}
\seealso{
Other Spoken LSAMon: 
\code{\link{LSAMonParse}()},
\code{\link{to_LSAMon_R_command}()}
}
\concept{Spoken LSAMon}
