% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-LSAMon.R
\name{LSAMonParse}
\alias{LSAMonParse}
\title{Parse LSAMon natural speech command.}
\usage{
LSAMonParse(command)
}
\arguments{
\item{command}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken LSAMon: 
\code{\link{LSAMonInterpret}()},
\code{\link{to_LSAMon_R_command}()}
}
\concept{Spoken LSAMon}
