% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLWebServiceFunctions.R
\name{MakeDSLWebServiceURL}
\alias{MakeDSLWebServiceURL}
\title{Make DSL web service URL}
\usage{
MakeDSLWebServiceURL(
  command = NULL,
  scheme = "http",
  hostname = "accendodata.net",
  port = "5040",
  path = "translate",
  query = NULL
)
}
\arguments{
\item{command}{The command to be interpreted.}

\item{scheme}{Scheme, string.}

\item{hostname}{Hostname, string.}

\item{port}{Port, string or integer.}

\item{path}{Path, string.}

\item{query}{A list with named elements of URL query specs.}
}
\value{
A URL string.
}
\description{
Gives interpreter web service URL of a DSL command.
}
\details{

}
\seealso{
Other DSLWebService: 
\code{\link{DSLWebServiceInterpretationURL}()},
\code{\link{DSLWebServiceInterpretation}()},
\code{\link{InterpretByDSLWebService}()}
}
\concept{DSLWebService}
