% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perl6-command-functions.R
\name{Perl6Command}
\alias{Perl6Command}
\title{Raku Perl 6 command}
\usage{
Perl6Command(
  command,
  moduleDirectory,
  moduleName,
  perl6Location = "/Applications/Rakudo/bin/raku"
)
}
\arguments{
\item{command}{A command string to be parsed.}

\item{moduleDirectory}{The directory name for the parsing module.}

\item{moduleName}{The name of the module to be loaded.}

\item{perl6Location}{Location of (path to) perl6..}
}
\value{
A string.
}
\description{
Parses a string with a given Raku Perl 6 module directory and
name, grammar, and grammar actions.
}
\details{
Most likely this function is going to be used inside a function with a simpler signature.
}
