% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perl6-command-functions.R
\name{Perl6Parse}
\alias{Perl6Parse}
\title{Raku Perl 6 parsing}
\usage{
Perl6Parse(
  command,
  moduleDirectory,
  moduleName,
  grammarClassName,
  actionsClassName = NULL,
  perl6Location = "/Applications/Rakudo/bin/perl6"
)
}
\arguments{
\item{command}{A command string to be parsed.}

\item{moduleDirectory}{The directory name for the parsing module.}

\item{grammarClassName}{The name of the grammar class.}

\item{actionsClassName}{The name of the actions class; if NULL just parsing is done.}

\item{perl6Location}{Location of (path to) perl6.}
}
\value{
A string
}
\description{
Parses a string with a given Raku Perl 6 module directory, module name,
grammar class, and grammar actions class.
}
\details{
Most likely this function is going to be used inside a function with a simpler signature.
}
