% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Parsing-and-interpretation-functions-hook-up-for-spoken-dplyr.R
\name{Perl6ParsingLib}
\alias{Perl6ParsingLib}
\title{Raku Perl 6 parser library}
\usage{
Perl6ParsingLib()
}
\value{
A string
}
\description{
Raku Perl 6 parser library
}
