% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-QRMon.R
\name{QRMonInterpret}
\alias{QRMonInterpret}
\title{Interpret QRMon natural speech command.}
\usage{
QRMonInterpret(command)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses and interprets a command by directly invoking a pair of Raku Perl6 parser and actions classes.
}
\seealso{
Other Spoken QRMon: 
\code{\link{QRMonParse}()},
\code{\link{to_QRMon_R_command}()}
}
\concept{Spoken QRMon}
