% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-QRMon.R
\name{QRMonParse}
\alias{QRMonParse}
\title{Parse QRMon natural speech command.}
\usage{
QRMonParse(command)
}
\arguments{
\item{command}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken QRMon: 
\code{\link{QRMonInterpret}()},
\code{\link{to_QRMon_R_command}()}
}
\concept{Spoken QRMon}
