% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RakuCommandFunctions.R
\name{RakuCommand}
\alias{RakuCommand}
\title{Raku Perl 6 command}
\usage{
RakuCommand(
  command,
  moduleDirectory,
  moduleName,
  rakuLocation = "/Applications/Rakudo/bin/raku"
)
}
\arguments{
\item{command}{A command string to be parsed.}

\item{moduleDirectory}{The directory name for the parsing module.}

\item{moduleName}{The name of the module to be loaded.}

\item{rakuLocation}{Location of (path to) raku..}
}
\value{
A string.
}
\description{
Parses a string with a given Raku Perl 6 module directory and
name, grammar, and grammar actions.
}
\details{
Most likely this function is going to be used inside a function with a simpler signature.
}
