% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RakuCommandFunctions.R
\name{RakuParse}
\alias{RakuParse}
\title{Raku Perl 6 parsing}
\usage{
RakuParse(
  command,
  moduleName,
  grammarClassName,
  actionsClassName = NULL,
  moduleDirectory = NULL,
  rakuLocation = "/Applications/Rakudo/bin/raku"
)
}
\arguments{
\item{command}{A command string to be parsed.}

\item{grammarClassName}{The name of the grammar class.}

\item{actionsClassName}{The name of the actions class; if NULL just parsing is done.}

\item{moduleDirectory}{The directory name for the parsing module.}

\item{rakuLocation}{Location of (path to) raku.}
}
\value{
A string
}
\description{
Parses a string with a given Raku Perl 6 module directory, module name,
grammar class, and grammar actions class.
}
\details{
Most likely this function is going to be used inside a function with a simpler signature.
}
