% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-SMRMon.R
\name{SMRMonInterpret}
\alias{SMRMonInterpret}
\title{Interpret SMRMon natural speech command.}
\usage{
SMRMonInterpret(command)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses and interprets a command by directly invoking a pair of Raku Perl6 parser and actions classes.
}
\seealso{
Other Spoken SMRMon: 
\code{\link{SMRMonParse}()},
\code{\link{to_SMRMon_R_command}()}
}
\concept{Spoken SMRMon}
