% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-SMRMon.R
\name{SMRMonParse}
\alias{SMRMonParse}
\title{Parse SMRMon natural speech command.}
\usage{
SMRMonParse(command)
}
\arguments{
\item{command}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken SMRMon: 
\code{\link{SMRMonInterpret}()},
\code{\link{to_SMRMon_R_command}()}
}
\concept{Spoken SMRMon}
