% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-QRMon.R
\name{to_QRMon_R_command}
\alias{to_QRMon_R_command}
\title{Interpret a QRMon natural language command.}
\usage{
to_QRMon_R_command(command, parse = TRUE)
}
\arguments{
\item{command}{A string with a command or a list of commands separated with ";".}

\item{parse}{A boolean should the result be parsed as an R expression.}
}
\value{
A string or an R expression
}
\description{
Calls Raku Perl 6 module function `to_QRMon_R_command` in order to get
interpretation of a natural language command or a list spoken commands separated with ";".
}
\details{
Produces a character vector or an expression depending on \code{parse}.
}
\seealso{
Other Spoken QRMon: 
\code{\link{QRMonInterpret}()},
\code{\link{QRMonParse}()}
}
\concept{Spoken QRMon}
