% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-dplyr.R
\name{DPLYRInterpret}
\alias{DPLYRInterpret}
\title{Interpret dplyr natural speech command.}
\usage{
DPLYRInterpret(command)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses and interprets a command by directly invoking a pair of Raku Perl6 parser and actions classes.
}
\seealso{
Other Spoken dplyr: 
\code{\link{DPLYRParse}()},
\code{\link{to_dplyr_command}()}
}
\concept{Spoken dplyr}
