% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Parsing-and-interpretation-functions-hook-up-for-spoken-dplyr.R
\name{DPLYRParse}
\alias{DPLYRParse}
\title{Parse dplyr natural speech command.}
\usage{
DPLYRParse(cmd)
}
\arguments{
\item{cmd}{A natural language command.}
}
\value{
A string
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken dplyr: \code{\link{DPLYRInterpret}},
  \code{\link{to_dplyr_command}}
}
\concept{Spoken dplyr}
