% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hook-up-for-spoken-dplyr.R
\name{DPLYRParse}
\alias{DPLYRParse}
\title{Parse dplyr natural speech command.}
\usage{
DPLYRParse(command)
}
\arguments{
\item{command}{A natural language command.}
}
\value{
A character vector
}
\description{
Parses a command by directly invoking a Raku Perl6 parser class.
}
\seealso{
Other Spoken dplyr: \code{\link{DPLYRInterpret}},
  \code{\link{SMRMonInterpret}}, \code{\link{SMRMonParse}},
  \code{\link{to_SMRMon_R_command}},
  \code{\link{to_dplyr_command}}
}
\concept{Spoken dplyr}
