% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSLWebServiceFunctions.R
\name{DSLWebServiceInterpretationURL}
\alias{DSLWebServiceInterpretationURL}
\title{DSL interpretation URL}
\usage{
DSLWebServiceInterpretationURL(
  command = NULL,
  query = NULL,
  url = "http://accendodata.net:5040/translate/",
  sub = NULL
)
}
\arguments{
\item{command}{The command to be interpreted.}

\item{query}{A character vector with named elements.}

\item{url}{The web service URL.}

\item{sub}{Sub-service name.
If NULL then only the \code{url} is used.}
}
\description{
Gives interpreter web service URL of a DSL command.
}
\seealso{
Other DSLWebService: 
\code{\link{DSLWebServiceInterpretation}()},
\code{\link{InterpretByDSLWebService}()},
\code{\link{MakeDSLWebServiceURL}()}
}
\concept{DSLWebService}
