% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Parsing-and-interpretation-functions-hook-up-for-spoken-dplyr.R
\name{LocalUserDirName}
\alias{LocalUserDirName}
\title{Local user directory name}
\usage{
LocalUserDirName()
}
\value{
A string
}
\description{
Local user directory name
}
