% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perl6-command-functions.R
\name{LocalUserDirName}
\alias{LocalUserDirName}
\title{Local user directory name}
\usage{
LocalUserDirName()
}
\value{
A string
}
\description{
Local user directory name
}
\details{
Mac OS X centric user directory at this point.
}
