% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Parsing-and-interpretation-functions-hook-up-for-spoken-dplyr.R
\docType{data}
\name{perl6ParsingLib}
\alias{perl6ParsingLib}
\title{Raku Perl 6 parser library}
\format{An object of class \code{character} of length 1.}
\usage{
perl6ParsingLib
}
\description{
Raku Perl 6 parser library
}
\keyword{datasets}
